+++ tests/Integration/Controller/ReviewControllerTest.php
@@ -12,8 +12,9 @@ namespace OxidEsales\GdprOptinModule\Tests\Integration\Controller;
 use OxidEsales\Eshop\Application\Controller\ReviewController;
 use OxidEsales\EshopCommunity\Tests\Integration\IntegrationTestCase;
 use OxidEsales\GdprOptinModule\Service\ReviewOptInInterface;
+use PHPUnit\Framework\Attributes\CoversClass;
 
-/** @covers \OxidEsales\GdprOptinModule\Controller\ReviewController */
+#[CoversClass(\OxidEsales\GdprOptinModule\Controller\ReviewController::class)]
 class ReviewControllerTest extends IntegrationTestCase
 {
     public function testInit(): void