+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 - Update module to work with OXID eShop 7.6
+- Upgrade the test suite to PHPUnit 12.5
 
 ### Fixed
 - Add file extension to exported files